Refactor EditTaskDialog component and introduce modular form handling
- Decomposed the EditTaskDialog into smaller components: EditTaskFormFields and EditTaskDialogActions for better organization and reusability.
- Implemented a custom hook, useEditTaskForm, to manage form state and submission logic, enhancing code clarity and separation of concerns.
- Updated the dialog to utilize the new components and hook, streamlining the task editing process.
